fix import json from v1 generator
