+++ sources/components/App.js
@@ -1,5 +1,6 @@
 // Main app component
-import { state, syncSelectionsToHash } from '../state/state.js';
+import { state, } from '../state/state.js';
+import { syncSelectionsToHash } from '../state/hash.js';
 import { Download } from './download/Download.js';
 import { FiltersPanel } from './FiltersPanel.js';
 import { Credits } from './download/Credits.js';
