+++ sources/components/download/Download.js
@@ -1,5 +1,6 @@
 // Download component
 import { state } from '../../state/state.js';
+import { extractAnimationFromCanvas, renderSingleItem, renderSingleItemAnimation, layers } from '../../canvas/renderer.js';
 import { getAllCredits, creditsToCsv, creditsToTxt, getItemFileName } from '../../utils/credits.js';
 import { CollapsibleSection } from '../CollapsibleSection.js';
 import { downloadFile, downloadAsPNG } from '../../canvas/download.js';
@@ -11,8 +12,9 @@ export const Download = {
 		// Export to clipboard
 		const exportToClipboard = async () => {
 			if (!window.canvasRenderer) return;
-			const json = exportStateAsJSON(state.selections, state.bodyType);
 			try {
+				const json = exportStateAsJSON(state, layers);
+				if (window.DEBUG) console.log(json);
 				await navigator.clipboard.writeText(json);
 				alert('Exported to clipboard!');
 			} catch (err) {
@@ -27,8 +29,7 @@ export const Download = {
 			try {
 				const json = await navigator.clipboard.readText();
 				const imported = importStateFromJSON(json);
-				state.bodyType = imported.bodyType;
-				state.selections = imported.selections;
+				Object.assign(state, imported);
 
 				m.redraw(); // Force Mithril to update the UI
 				alert('Imported successfully!');
@@ -71,7 +72,7 @@ export const Download = {
 				// Create animation PNGs in standard folder (no custom animations support yet)
 				for (const anim of animationList) {
 					try {
-						const animCanvas = window.canvasRenderer.extractAnimationFromCanvas(anim.value);
+						const animCanvas = extractAnimationFromCanvas(anim.value);
 						if (animCanvas) {
 							const blob = await new Promise(resolve => animCanvas.toBlob(resolve, 'image/png'));
 							standardFolder.file(`${anim.value}.png`, blob);
@@ -154,7 +155,7 @@ export const Download = {
 
 					try {
 						// Render just this one item
-						const itemCanvas = await window.canvasRenderer.renderSingleItem(
+						const itemCanvas = await renderSingleItem(
 							itemId,
 							variant,
 							bodyType,
@@ -237,7 +238,7 @@ export const Download = {
 
 						try {
 							// Render just this item for this animation
-							const animCanvas = await window.canvasRenderer.renderSingleItemAnimation(
+							const animCanvas = await renderSingleItemAnimation(
 								itemId,
 								variant,
 								bodyType,
