+++ sources/state/path.js
@@ -1,4 +1,4 @@
-import { getHashParamsforSelections } from './state.js';
+import { getHashParamsforSelections } from './hash.js';
 import { variantToFilename, es6DynamicTemplate } from '../utils/helpers.js';
 
 /**
